requirements.txt
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
connexion==2.14.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.44
swagger-ui-bundle==0.0.9
typing_extensions>=4.15.0
urllib3==2.0.6
Werkzeug==2.2.2
pyodbc
rpds-py>=0.13.2
pytz
uvicorn==0.38.0
fastapi==0.124.0
python-multipart==0.0.20
PyJWT>=2.8.0